[threading] Change request async suspend to return 3 states instead of one to account for async racing to the middle of a self suspend.
When begin async suspend lands in STATE_SUSPEND_IN_PROGRESS, the initiator don't need to perform an async suspend, it just needs to wait.

The alternative would be to change finish async suspend to account for the possibility of witnessing STATE_SUSPEND_PROMOTED_TO_ASYNC.

I went with the first option as it requires less async suspend and doesn't increase the valid state space.